requirements.txt
Flask
pytesseract
Pillow
gunicorn